# api/docs/swagger.yaml
definitions:
...
  auth_service.HasAccessRequest:
    properties:
      access_token:
        type: string
      client_platform_id:
        type: string
      method:
        type: string
      path:
        type: string
      project_id:
        type: string
    type: object
  auth_service.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  auth_service.LogoutRequest:
    properties:
      access_token:
        type: string
    type: object
...
  auth_service.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  auth_service.User:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
    type: object
...
  http.Response:
    properties:
      data:
        type: object
      description:
        type: string
      status:
        type: string
    type: object
  settings_service.AddLevelItemRequest:
    properties:
      description:
        type: string
      position_id:
        type: string
      title:
        type: string
    type: object
  settings_service.CreatePositionRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
...
  settings_service.GetPositionListResponse:
    properties:
      count:
        type: integer
      positions:
        items:
          $ref: '#/definitions/settings_service.PositionSlim'
        type: array
    type: object
...
  settings_service.Position:
    properties:
      description:
        type: string
      id:
        type: string
      level_items:
        items:
          $ref: '#/definitions/settings_service.Position_LevelItem'
        type: array
      title:
        type: string
    type: object
  settings_service.Position_LevelItem:
    properties:
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  settings_service.PositionSlim:
    properties:
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
  settings_service.UpdateLevelItemRequest:
    properties:
      description:
        type: string
      id:
        type: string
      position_id:
        type: string
      title:
        type: string
    type: object
  settings_service.UpdatePositionRequest:
    properties:
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  license: {}
  termsOfService: https://udevs.io
paths:
...
  /has-access:
    post:
      consumes:
      - application/json
      description: Has Access
      operationId: has_access
      parameters:
      - description: HasAccessRequestBody
        in: body
        name: has-access
        required: true
        schema:
          $ref: '#/definitions/auth_service.HasAccessRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Has Access
      tags:
      - Session
  /login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: login
      parameters:
      - description: LoginRequestBody
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth_service.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login
      tags:
      - Session
  /logout:
    delete:
      consumes:
      - application/json
      description: Logout User
      operationId: logout
      parameters:
      - description: LogoutRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth_service.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Logout User
      tags:
      - Session
...
  /position:
    get:
      consumes:
      - application/json
      description: Get Position List
      operationId: get_position_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetPositionListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/settings_service.GetPositionListResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Position List
      tags:
      - Position
    post:
      consumes:
      - application/json
      description: Create Position
      operationId: create_position
      parameters:
      - description: position
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/settings_service.CreatePositionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: PositionBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/settings_service.Position'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Position
      tags:
      - Position
    put:
      consumes:
      - application/json
      description: Update Position
      operationId: update_position
      parameters:
      - description: position
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/settings_service.UpdatePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PositionBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/settings_service.Position'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Position
      tags:
      - Position
  /position/{position-id}:
    delete:
      consumes:
      - application/json
      description: Get Position
      operationId: delete_position
      parameters:
      - description: position-id
        in: path
        name: position-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Position
      tags:
      - Position
    get:
      consumes:
      - application/json
      description: Get Position By ID
      operationId: get_position_by_id
      parameters:
      - description: position-id
        in: path
        name: position-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PositionBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/settings_service.Position'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Position By ID
      tags:
      - Position
  /position/{position-id}/level/{level-id}:
    delete:
      consumes:
      - application/json
      description: Remove Level Item
      operationId: remove_level_item
      parameters:
      - description: position-id
        in: path
        name: position-id
        required: true
        type: string
      - description: level-id
        in: path
        name: level-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Remove Level Item
      tags:
      - Position
  /position/level:
    post:
      consumes:
      - application/json
      description: Add Level Item
      operationId: add_level_item
      parameters:
      - description: AddLevelItemRequestBody
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/settings_service.AddLevelItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: PositionBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/settings_service.Position'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Add Level Item
      tags:
      - Position
    put:
      consumes:
      - application/json
      description: Update Level Item
      operationId: update_level_item
      parameters:
      - description: UpdateLevelItemRequestBody
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/settings_service.UpdateLevelItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PositionBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/settings_service.Position'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Level Item
      tags:
      - Position
  /refresh:
    put:
      consumes:
      - application/json
      description: Refresh Token
      operationId: refresh
      parameters:
      - description: RefreshTokenRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Refresh Token
      tags:
      - Session
...
swagger: "2.0"